Tabla de postulaciones en perfil de empresa falta filtrar los datos por publicación
